# Fix distributed tracing silent failures and implement comprehensive tracing architecture