--- a/observability/grafana/grafana.ini
+++ b/observability/grafana/grafana.ini
@@ -10,7 +10,7 @@ disable_gravatar = true
 [users]
 allow_sign_up = false
 allow_org_create = false
-home_page = /d/mcp-mesh-overview/mcp-mesh-overview
+home_page = /grafana/d/mcp-mesh-overview/mcp-mesh-overview
 default_theme = dark
 
 [auth.anonymous]
@@ -29,7 +29,9 @@ provisioning = /etc/grafana/provisioning
 [server]
 http_port = 3000
 domain = localhost
-root_url = http://localhost:3000/
+root_url = /grafana/
+serve_from_sub_path = true
+enforce_domain = false
 
 [database]
 type = sqlite3