--- a/src/runtime/python/_mcp_mesh/engine/dependency_injector.py
+++ b/src/runtime/python/_mcp_mesh/engine/dependency_injector.py
@@ -2,6 +2,8 @@
 Dynamic dependency injection system for MCP Mesh.
 
 Handles both initial injection and runtime updates when topology changes.
+Focused purely on dependency injection - telemetry/tracing is handled at 
+the HTTP middleware layer for unified approach across MCP agents and FastAPI apps.
 """
 
 import asyncio
@@ -292,41 +294,23 @@ class DependencyInjector:
             if inspect.iscoroutinefunction(func):
                 @functools.wraps(func)
                 async def minimal_wrapper(*args, **kwargs):
-                    # Execute with telemetry tracing even for async functions without dependencies
-                    try:
-                        from ..tracing.execution_tracer import ExecutionTracer
-                        
-                        # Use ExecutionTracer for minimal async wrapper (no dependencies)
-                        return await ExecutionTracer.trace_function_execution_async(
-                            func, args, kwargs, [], [], 0, wrapper_logger
-                        )
-                    except ImportError:
-                        # Fallback if tracing is unavailable - never fail user function
-                        wrapper_logger.debug("🔇 Tracing unavailable, executing without telemetry")
-                        return await func(*args, **kwargs)
-                    except Exception as e:
-                        # Never fail user function due to tracing errors
-                        wrapper_logger.warning(f"⚠️ Telemetry failed, executing without: {e}")
-                        return await func(*args, **kwargs)
+                    # Use ExecutionTracer for functions without dependencies (v0.4.0 style)
+                    from ..tracing.execution_tracer import ExecutionTracer
+                    wrapper_logger.debug(f"🔧 DI: Executing async function {func.__name__} (no dependencies)")
+                    
+                    # For async functions without dependencies, use the async tracer
+                    return await ExecutionTracer.trace_function_execution_async(
+                        func, args, kwargs, [], [], 0, wrapper_logger
+                    )
             else:
                 @functools.wraps(func)
                 def minimal_wrapper(*args, **kwargs):
-                    # Execute with telemetry tracing even for functions without dependencies
-                    try:
-                        from ..tracing.execution_tracer import ExecutionTracer
-                        
-                        # Use ExecutionTracer for minimal wrapper (no dependencies)
-                        return ExecutionTracer.trace_original_function(
-                            func, args, kwargs, wrapper_logger
-                        )
-                    except ImportError:
-                        # Fallback if tracing is unavailable - never fail user function
-                        wrapper_logger.debug("🔇 Tracing unavailable, executing without telemetry")
-                        return func(*args, **kwargs)
-                    except Exception as e:
-                        # Never fail user function due to tracing errors
-                        wrapper_logger.warning(f"⚠️ Telemetry failed, executing without: {e}")
-                        return func(*args, **kwargs)
+                    # Use ExecutionTracer for functions without dependencies (v0.4.0 style)
+                    from ..tracing.execution_tracer import ExecutionTracer
+                    wrapper_logger.debug(f"🔧 DI: Executing sync function {func.__name__} (no dependencies)")
+                    
+                    # Use original function tracer for functions without dependencies
+                    return ExecutionTracer.trace_original_function(func, args, kwargs, wrapper_logger)
 
             # Add minimal metadata for compatibility
             minimal_wrapper._mesh_injected_deps = {}
@@ -426,54 +410,29 @@ class DependencyInjector:
                     f"🔧 DEPENDENCY_WRAPPER: final_kwargs={final_kwargs}"
                 )
 
-                # ===== EXECUTE WITH TELEMETRY AND DEPENDENCY INJECTION =====
-                # Call the original function with telemetry tracing and injected dependencies
+                # ===== EXECUTE WITH DEPENDENCY INJECTION AND TRACING =====
+                # Use ExecutionTracer for comprehensive execution logging (v0.4.0 style)
+                from ..tracing.execution_tracer import ExecutionTracer
+                
                 original_func = func._mesh_original_func
 
-                # Execute with telemetry tracing
-                try:
-                    from ..tracing.execution_tracer import ExecutionTracer
-                    
-                    # Use ExecutionTracer for comprehensive telemetry
-                    if inspect.iscoroutinefunction(original_func):
-                        # For async functions, await the result directly
-                        result = await ExecutionTracer.trace_function_execution_async(
-                            original_func,
-                            args,
-                            final_kwargs,
-                            dependencies,
-                            mesh_positions,
-                            injected_count,
-                            wrapper_logger,
-                        )
-                    else:
-                        # For sync functions, call directly
-                        result = ExecutionTracer.trace_function_execution(
-                            original_func,
-                            args,
-                            final_kwargs,
-                            dependencies,
-                            mesh_positions,
-                            injected_count,
-                            wrapper_logger,
-                        )
-                except ImportError:
-                    # Fallback if tracing is unavailable - never fail user function
-                    wrapper_logger.debug("🔇 Tracing unavailable, executing without telemetry")
-                    if inspect.iscoroutinefunction(original_func):
-                        result = await original_func(*args, **final_kwargs)
-                    else:
-                        result = original_func(*args, **final_kwargs)
-                except Exception as e:
-                    # Never fail user function due to tracing errors
-                    wrapper_logger.warning(f"⚠️ Telemetry failed, executing without: {e}")
-                    if inspect.iscoroutinefunction(original_func):
-                        result = await original_func(*args, **final_kwargs)
-                    else:
-                        result = original_func(*args, **final_kwargs)
+                wrapper_logger.debug(
+                    f"🔧 DI: Executing async function {original_func.__name__} with {injected_count} injected dependencies"
+                )
+
+                # Use ExecutionTracer's async method for clean tracing
+                result = await ExecutionTracer.trace_function_execution_async(
+                    original_func,
+                    args,
+                    final_kwargs,
+                    dependencies,
+                    mesh_positions,
+                    injected_count,
+                    wrapper_logger,
+                )
 
                 wrapper_logger.debug(
-                    f"🔧 DEPENDENCY_WRAPPER: Original returned: {type(result)}"
+                    f"🔧 DI: Function {original_func.__name__} returned: {type(result)}"
                 )
                 return result
 
@@ -524,29 +483,24 @@ class DependencyInjector:
                             final_kwargs[param_name] = dependency
                             injected_count += 1
 
-                # ===== EXECUTE WITH TELEMETRY AND DEPENDENCY INJECTION =====
-                # Call the original function with telemetry tracing and injected dependencies
-                try:
-                    from ..tracing.execution_tracer import ExecutionTracer
-                    
-                    # Use ExecutionTracer for comprehensive telemetry
-                    return ExecutionTracer.trace_function_execution(
-                        func._mesh_original_func,
-                        args,
-                        final_kwargs,
-                        dependencies,
-                        mesh_positions,
-                        injected_count,
-                        wrapper_logger,
-                    )
-                except ImportError:
-                    # Fallback if tracing is unavailable - never fail user function
-                    wrapper_logger.debug("🔇 Tracing unavailable, executing without telemetry")
-                    return func._mesh_original_func(*args, **final_kwargs)
-                except Exception as e:
-                    # Never fail user function due to tracing errors
-                    wrapper_logger.warning(f"⚠️ Telemetry failed, executing without: {e}")
-                    return func._mesh_original_func(*args, **final_kwargs)
+                # ===== EXECUTE WITH DEPENDENCY INJECTION AND TRACING =====
+                # Use ExecutionTracer for comprehensive execution logging (v0.4.0 style)
+                from ..tracing.execution_tracer import ExecutionTracer
+                
+                wrapper_logger.debug(
+                    f"🔧 DI: Executing sync function {func._mesh_original_func.__name__} with {injected_count} injected dependencies"
+                )
+
+                # Use ExecutionTracer for clean execution tracing
+                return ExecutionTracer.trace_function_execution(
+                    func._mesh_original_func,
+                    args,
+                    final_kwargs,
+                    dependencies,
+                    mesh_positions,
+                    injected_count,
+                    wrapper_logger,
+                )
 
         # Store dependency state on wrapper
         dependency_wrapper._mesh_injected_deps = {}